+++ Runtime/DisableInInspector.cs
@@ -3,5 +3,5 @@ using System;
 namespace Kiryonn.UnityAttributes
 {
 	[AttributeUsage(AttributeTargets.Field)]
-	public class DisableInInspector : UnityEngine.PropertyAttribute { }
+	public class DisableInInspectorAttribute : UnityEngine.PropertyAttribute { }
 }