+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [1-0-0] - 2025-06-06
-### First Release
-- Added DisableInInspector to display a field in the inspector but not let user interact with it.
+## [1.0.1]
+### Changes
+- Now use [Kiryonn Searchables](https://github.com/Kiryonn/com.kiryonn.unity-searchables) for [SerializeAbstractField](Runtime/SerializeAbstractField.cs) type chosing.
+- Refactored [SerializeAbstractField](Runtime/SerializeAbstractField.cs) to look better and work faster.
+
+### BugFix
+- [SerializeAbstractField](Runtime/SerializeAbstractField.cs) now correctly interact with multiple and nested Generic Type Arguments.
+
+
+## [1.0.0] First Release
+- Added [DisableInInspector](Runtime/DisableInInspector.cs) and [SerializeAbstractField](Runtime/SerializeAbstractField.cs)